Update provider spec & upgrade to llama-stack 0.2.22
- Split into inline & remote modules.
- Update `run-inline.yaml` and `run-remote.yaml` to include new APIs and adjust paths.
- Remove deprecated provider YAML files and implement new inline and remote provider structures.
- Modify inline and remote demo notebooks to reflect updated model identifiers and URLs.
