+++ distribution/run-inline.yaml
@@ -5,10 +5,13 @@ apis:
   - inference
   - telemetry
   - datasetio
+  - files
+  - benchmarks
 providers:
   eval:
-    - provider_id: trustyai_ragas_inline
+    - provider_id: trustyai_ragas
       provider_type: inline::trustyai_ragas
+      module: llama_stack_provider_ragas.inline
       config:
         embedding_model: ${env.EMBEDDING_MODEL}
   datasetio:
@@ -18,19 +21,29 @@ providers:
         kvstore:
           type: sqlite
           namespace: null
-          db_path: ${env.SQLITE_STORE_DIR:~/.llama/distributions/trustyai_ragas_distro}/localfs_datasetio.db
+          db_path: ${env.SQLITE_STORE_DIR:=~/.llama/distributions/trustyai_ragas_distro_inline}/localfs_datasetio.db
   inference:
     - provider_id: ollama
       provider_type: remote::ollama
       config:
-        url: ${env.OLLAMA_URL:http://localhost:11434}
+        url: ${env.OLLAMA_URL:=http://localhost:11434}
   telemetry:
     - provider_id: meta-reference
       provider_type: inline::meta-reference
       config:
-        service_name: ${env.OTEL_SERVICE_NAME:}
-      sinks: ${env.TELEMETRY_SINKS:console,sqlite}
-      sqlite_db_path: ${env.SQLITE_STORE_DIR:~/.llama/distributions/trustyai_ragas_distro}/trace_store.db
+        service_name: "${env.OTEL_SERVICE_NAME:=\u200B}"
+        sinks: ${env.TELEMETRY_SINKS:=console,sqlite}
+        sqlite_db_path: ${env.SQLITE_STORE_DIR:=~/.llama/distributions/trustyai_ragas_distro_inline}/trace_store.db
+        otel_exporter_otlp_endpoint: ${env.OTEL_EXPORTER_OTLP_ENDPOINT:=}
+  files:
+    - provider_id: meta-reference-files
+      provider_type: inline::localfs
+      config:
+        storage_dir: ${env.FILES_STORAGE_DIR:=~/.llama/distributions/trustyai_ragas_distro_inline/files}
+        metadata_store:
+          type: sqlite
+          db_path: ${env.METADATA_STORE_DB_PATH:=~/.llama/distributions/trustyai_ragas_distro_inline}/registry.db}
+
 models:
   - metadata:
       embedding_dimension: 384
@@ -46,4 +59,3 @@ models:
 server:
   host: localhost
   port: 8321
-external_providers_dir: ./providers.d
