+++ distribution/run-remote.yaml
@@ -3,12 +3,15 @@ image_name: trustyai_ragas_distro_remote
 apis:
   - eval
   - inference
+  - files
+  - benchmarks
   - telemetry
   - datasetio
 providers:
   eval:
-    - provider_id: trustyai_ragas_remote
+    - provider_id: trustyai_ragas
       provider_type: remote::trustyai_ragas
+      module: llama_stack_provider_ragas.remote
       config:
         embedding_model: ${env.EMBEDDING_MODEL}
         kubeflow_config:
@@ -25,19 +28,28 @@ providers:
         kvstore:
           type: sqlite
           namespace: null
-          db_path: ${env.SQLITE_STORE_DIR:~/.llama/distributions/trustyai_ragas_distro}/localfs_datasetio.db
+          db_path: ${env.SQLITE_STORE_DIR:=~/.llama/distributions/trustyai_ragas_distro_remote}/localfs_datasetio.db
   inference:
     - provider_id: ollama
       provider_type: remote::ollama
       config:
-        url: ${env.OLLAMA_URL:http://localhost:11434}
+        url: ${env.OLLAMA_URL:=http://localhost:11434}
   telemetry:
     - provider_id: meta-reference
       provider_type: inline::meta-reference
       config:
-        service_name: ${env.OTEL_SERVICE_NAME:}
-      sinks: ${env.TELEMETRY_SINKS:console,sqlite}
-      sqlite_db_path: ${env.SQLITE_STORE_DIR:~/.llama/distributions/trustyai_ragas_distro}/trace_store.db
+        service_name: "${env.OTEL_SERVICE_NAME:=\u200B}"
+        sinks: ${env.TELEMETRY_SINKS:=console,sqlite}
+        sqlite_db_path: ${env.SQLITE_STORE_DIR:=~/.llama/distributions/trustyai_ragas_distro_remote}/trace_store.db
+        otel_exporter_otlp_endpoint: ${env.OTEL_EXPORTER_OTLP_ENDPOINT:=}
+  files:
+    - provider_id: meta-reference-files
+      provider_type: inline::localfs
+      config:
+        storage_dir: ${env.FILES_STORAGE_DIR:=~/.llama/distributions/trustyai_ragas_distro_remote/files}
+        metadata_store:
+          type: sqlite
+          db_path: ${env.METADATA_STORE_DB_PATH:=~/.llama/distributions/trustyai_ragas_distro_remote}/registry.db}
 models:
   - metadata:
       embedding_dimension: 384
@@ -53,4 +65,3 @@ models:
 server:
   host: localhost
   port: 8321
-external_providers_dir: ./providers.d
