--- providers.d/inline/eval/trustyai_ragas.yaml
@@ -1,5 +1,0 @@
-module: llama_stack_provider_ragas
-config_class: llama_stack_provider_ragas.config.RagasProviderInlineConfig
-pip_packages: ["ragas"] #, ["datasets", "langchain-core"] # TODO: i think i need to add these
-api_dependencies: ["inference", "datasetio"]
-optional_api_dependencies: []
