+++ src/llama_stack_provider_ragas/remote/kubeflow/components.py
@@ -42,7 +42,7 @@ def run_ragas_evaluation(
 
     from llama_stack_provider_ragas.constants import METRIC_MAPPING
     from llama_stack_provider_ragas.logging_utils import render_dataframe_as_table
-    from llama_stack_provider_ragas.wrappers_remote import (
+    from llama_stack_provider_ragas.remote.wrappers_remote import (
         LlamaStackRemoteEmbeddings,
         LlamaStackRemoteLLM,
     )
